Separate distribution logic
